fix: solve reloading after config changes
