+++ src/main.ts
@@ -1,0 +1,18 @@
+import { workspace, window } from "vscode";
+import { getThemePaths } from "./helpers";
+import utils from "./utils";
+
+export const activate = () => {
+  workspace.onDidChangeConfiguration((event) => {
+    if (event.affectsConfiguration("catppuccin")) {
+      const config = utils.getConfiguration();
+      const paths = getThemePaths();
+      utils.updateThemes(config, paths);
+      window.showInformationMessage(
+        "Catppuccin Configuration changed, re-generating the JSON theme."
+      );
+    }
+  });
+};
+
+export const deactivate = () => {};
